loading/lazyload/image-loading-lazy-out-of-viewport-dynamic.tentative.html

<!DOCTYPE html>
<head>
  <title>Out of viewport images with loading='lazy' and changed to
         loading='eager' load and do not block the window load event</title>
  <link rel="author" title="Rob Buis" href="mailto:rbuis@igalia.com">
  <link rel="help" href="https://github.com/scott-little/lazyload">
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>

<!--
Marked as tentative until https://github.com/whatwg/html/pull/3752 is landed.
-->

<script>
  const t = async_test("Test that out of viewport images with loading='lazy' " +
                       "and changed to loading='eager' load and do not block " +
                       "the window load event.");

  let has_window_loaded = false;

  window.addEventListener("load", t.step_func(function() {
    assert_false(has_window_loaded,
                 "The window.load() event should only fire once.");
    has_window_loaded = true;
  }));

  const below_viewport_img_onload = t.step_func_done(function() {
    assert_true(has_window_loaded,
                "The window.load() event should have fired before " +
                "below_viewport loaded.");
  });
</script>

<body>
  <div style="height:10000px;"></div>
  <img id="below_viewport" src="resources/image.png?pipe=trickle(d2)"
       loading="lazy" onload="below_viewport_img_onload();">
  <script>
    document.getElementById("below_viewport").loading = 'eager';
  </script>
</body>
